requirements.txt
python-docx==0.8.11
openpyxl==3.10.10
python-pptx==0.6.23
PyPDF2==3.0.1
google-auth-oauthlib==1.2.1
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth==2.29.0
